Fix ckeditor5-image config imports
