+++ packages/ckeditor5-image/src/imageinsert.ts
@@ -40,81 +40,9 @@ export default class ImageInsert extends Plugin {
 	}
 }
 
-/**
- * The configuration of the image insert dropdown panel view. Used by the image insert feature in the `@ckeditor/ckeditor5-image` package.
- *
- * ```ts
- * ClassicEditor
- * 	.create( editorElement, {
- * 			image: {
- * 				insert: {
- * 				... // settings for "insertImage" view goes here
- * 				}
- * 			}
- * 	} )
- * 	.then( ... )
- * 	.catch( ... );
- * ```
- *
- * See {@link module:core/editor/editorconfig~EditorConfig all editor options}.
- */
-export interface ImageInsertConfig {
-
-	/**
-	 * The image insert panel view configuration contains a list of {@link module:image/imageinsert~ImageInsert} integrations.
-	 *
-	 * The option accepts string tokens.
-	 * * for predefined integrations, we have two special strings: `insertImageViaUrl` and `openCKFinder`.
-	 * The former adds the **Insert image via URL** feature, while the latter adds the built-in **CKFinder** integration.
-	 * * for custom integrations, each string should be a name of the component registered in the
-	 * {@link module:ui/componentfactory~ComponentFactory component factory}.
-	 * If you have a plugin `PluginX` that registers `pluginXButton` component, then the integration token
-	 * in that case should be `pluginXButton`.
-	 *
-	 * ```ts
-	 * // Add `insertImageViaUrl`, `openCKFinder` and custom `pluginXButton` integrations.
-	 * const imageInsertConfig = {
-	 * 	insert: {
-	 * 		integrations: [
-	 * 			'insertImageViaUrl',
-	 * 			'openCKFinder',
-	 * 			'pluginXButton'
-	 * 		]
-	 * 	}
-	 * };
-	 * ```
-	 *
-	 * @internal
-	 * @default [ 'insertImageViaUrl' ]
-	 */
-	integrations: Array<string>;
-
-	/**
-	 * This options allows to override the image type used by the {@link module:image/image/insertimagecommand~InsertImageCommand}
-	 * when the user inserts new images into the editor content. By default, this option is unset which means the editor will choose
-	 * the optimal image type based on the context of the insertion (e.g. the current selection and availability of plugins)
-	 *
-	 * Available options are:
-	 *
-	 * * `'block'` – all images inserted into the editor will be block (requires the {@link module:image/imageblock~ImageBlock} plugin),
-	 * * `'inline'` – all images inserted into the editor will be inline (requires the {@link module:image/imageinline~ImageInline} plugin).
-	 */
-	type?: 'inline' | 'block';
-}
-
 declare module '@ckeditor/ckeditor5-core' {
 	interface PluginsMap {
 		[ ImageInsert.pluginName ]: ImageInsert;
-	}
-}
-
-declare module './image' {
-	interface ImageConfig {
-
-		/**
-		 * The image insert configuration.
-		 */
-		insert?: ImageInsertConfig;
 	}
 }
