+++ packages/ckeditor5-image/src/imageresize/imageresizeediting.ts
@@ -11,6 +11,7 @@ import type { ViewElement } from 'ckeditor5/src/engine';
 import { type Editor, Plugin, type PluginDependencies } from 'ckeditor5/src/core';
 import ImageUtils from '../imageutils';
 import ResizeImageCommand from './resizeimagecommand';
+import '../imageconfig';
 
 /**
  * The image resize editing feature.
