+++ packages/ckeditor5-image/src/imagestyle/imagestyleui.ts
@@ -12,7 +12,7 @@ import { ButtonView, createDropdown, addToolbarToDropdown, SplitButtonView } fro
 import { isObject, identity } from 'lodash-es';
 import ImageStyleEditing from './imagestyleediting';
 import utils from './utils';
-import type { ImageStyleOptionDefinition } from '../imagestyle';
+import type { ImageStyleDropdownDefinition, ImageStyleOptionDefinition } from '../imageconfig';
 
 import '../../theme/imagestyle.css';
 
@@ -241,68 +241,7 @@ function getDropdownButtonTitle( dropdownTitle: string | undefined, buttonTitle:
 	return ( dropdownTitle ? dropdownTitle + ': ' : '' ) + buttonTitle;
 }
 
-/**
- * # **The image style custom drop-down definition descriptor**
- *
- * This definition can be implemented in the {@link module:image/image~ImageConfig#toolbar image toolbar configuration}
- * to define a completely custom drop-down in the image toolbar.
- *
- * ```ts
- * ClassicEditor.create( editorElement, {
- * 	image: { toolbar: [
- * 			// One of the predefined drop-downs
- * 			'imageStyle:wrapText',
- * 		// Custom drop-down
- * 		{
- * 			name: 'imageStyle:customDropdown',
- * 			title: Custom drop-down title,
- * 			items: [ 'imageStyle:alignLeft', 'imageStyle:alignRight' ],
- * 			defaultItem: 'imageStyle:alignLeft'
- * 		}
- * 	] }
- * } );
- * ```
- *
- * **Note:** At the moment it is possible to populate the custom drop-down with only the buttons registered by the `ImageStyle` plugin.
- *
- * The defined drop-down will be registered
- * as the {@link module:ui/dropdown/dropdownview~DropdownView}
- * with the {@link module:ui/dropdown/button/splitbuttonview~SplitButtonView} under the provided name in the
- * {@link module:ui/componentfactory~ComponentFactory}
- */
-export interface ImageStyleDropdownDefinition {
-
-	/**
-	 * The unique name of the drop-down. It is recommended to precede it with the "imageStyle:" prefix
-	 * to avoid collision with the components' names registered by other plugins.
-	 */
-	name: string;
-
-	/**
-	 * The drop-down's title. It will be used as the split button label along with the title of the default item
-	 * in the following manner: "Custom drop-down title: Default item title".
-	 *
-	 * Setting `title` to one of
-	 * {@link module:image/imagestyle/imagestyleui~ImageStyleUI#localizedDefaultStylesTitles}
-	 * will automatically translate it to the language of the editor.
-	 */
-	title?: string;
-
-	/**
-	 * The list of the names of the buttons that will be placed in the drop-down's toolbar.
-	 * Each of the buttons has to be one of the {@link module:image/image~ImageConfig#styles default image style buttons}
-	 * or to be defined as the {@link module:image/imagestyle~ImageStyleOptionDefinition image styling option}.
-	 */
-	items: Array<string>;
-
-	/**
-	 * defaultItem The name of one of the buttons from the items list,
-	 * which will be used as a default button for the drop-down's split button.
-	 */
-	defaultItem: string;
- }
-
- declare module '@ckeditor/ckeditor5-core' {
+declare module '@ckeditor/ckeditor5-core' {
 	interface PluginsMap {
 		[ ImageStyleUI.pluginName ]: ImageStyleUI;
 	}
